pyinstaller and repo work

+++ build/build.ps1
@@ -6,9 +6,41 @@ param (
 )
 
 function clean() {
-    Remove-Item -Path $PSScriptRoot\build -Force -Recurse
-    Remove-Item -Path $PSScriptRoot\dist -Force -Recurse
-    Remove-Item -Path $PSScriptRoot\*.spec -Force
+    if (Test-Path -Path $PSScriptRoot\build) {
+        Remove-Item -Path $PSScriptRoot\build -Force -Recurse
+    }
+    if (Test-Path -Path $PSScriptRoot\dist) {
+        Remove-Item -Path $PSScriptRoot\dist -Force -Recurse
+    }
+    if (Test-Path -Path $PSScriptRoot\*.spec) {
+        Remove-Item -Path $PSScriptRoot\*.spec -Force
+    }
+    if (Test-Path -Path $PSScriptRoot\..\src\modules\app\__pycache__) {
+        Remove-Item -Path $PSScriptRoot\..\src\modules\app\__pycache__ -Force -Recurse
+    }
+    if (Test-Path -Path $PSScriptRoot\..\src\modules\gui\__pycache__) {
+        Remove-Item -Path $PSScriptRoot\..\src\modules\gui\__pycache__ -Force -Recurse
+    }
+}
+
+function remove-pyinstaller() {
+    . $PSScriptRoot\..\.venv\Scripts\Activate.ps1
+
+    pip uninstall altgraph pefile pyinstaller pyinstaller-hooks-contrib pywin32-ctypes -y
+}
+
+function requirements() {
+    # rebuilds the requirements.txt file in the project root
+    . $PSScriptRoot\..\.venv\Scripts\Activate.ps1
+
+    pip freeze --exclude setuptools `
+               --exclude pip `
+               --exclude pyinstaller `
+               --exclude altgraph `
+               --exclude pefile `
+               --exclude pyinstaller-hooks-contrib `
+               --exclude wheel `
+               --exclude pywin32-ctypes > $PSScriptRoot\..\requirements.txt
 }
 
 function build() {
@@ -23,10 +55,20 @@ function build() {
 }
 
 switch ($Command) {
-    build { build }
+    build {
+        requirements
+        build 
+    }
     clean { clean }
     cleanbuild {
         clean
+        requirements
         build
+    }
+    requirements {
+        requirements
+    }
+    remove-pyinstaller {
+        remove-pyinstaller
     }
 }
